18. 4Sum
